// src/app/modules/main/pages/main-meetings/models/meeting.model.ts
export interface IMeetingModel {
  applyStatusNameEn: string;
  createUserId: string;
  meetingStatus: string;
  conferenceTypeNameAz: string;
  orgNameRu: string;
  durationStatusName: string;
  moderator: string;
  discussion: any[];
  type: string;
  organizationId: string;
  conferenceType: string;
  typeNameRu: string;
  conferenceTypeNameEn: string;
  activityStatus: string;
  confirmStatus: string;
  userFullname: string;
  id: string;
  orgNameAz: string;
  moderatorFullname: string;
  orgFormula: string;
  orgName: string;
  updateUserId: string;
  active: string;
  agendaCount: string;
  orgNameEn: string;
  moderatorStatus: string;
  applyStatusName: string;
  applyStatus: string;
  startDate: string;
  note: string;
  updateDate: string;
  declineActivityStatus: string;
  durationStatus: string;
  typeName: string;
  meetingTypeName: string;
  title: string;
  meetingType: string;
  participantId: string;
  duration: string;
  meetingTypeNameRu: string;
  durationStatusNameEn: string;
  startTime: string;
  userEmail: string;
  declineReason: string;
  createDate: string;
  typeNameAz: string;
  recordFileId: string;
  decision: any[];
  durationStatusNameAz: string;
  meetingTypeNameAz: string;
  userId: string;
  conferenceTypeName: string;
  applyStatusNameAz: string;
  typeNameEn: string;
  pollCount: string;
  moderatorEmail: string;
  durationStatusNameRu: string;
  applyStatusNameRu: string;
  conferenceTypeNameRu: string;
  meetingTypeNameEn: string;
}